Make sure we set client_window/widget to null if app set it.
When set_client_window/widget passes null, we did not handle it
correctly. Usually this is not happen on regular gtk app like gedit so
the issue is not exposed. For example, gedit won't set client_window to
null when it lost focus, because in regular gtk widget, context is owned
by a certain gtk widget and only set when the gtk widget is destroyed.

In emacs pgtk, when it lost focus, set_client_window null will be called
but we failed to reset internal client_window also to null. So when it
regains focus again, next set_client_window (not null) is ignored by
client_window == context->client_window, and thus it doesn't work as expected.

Fix #19