Change GtkIMContext.reset to always commit preedit. (fcitx/fcitx5-unikey#35)
In Gtk2/Gtk3, also use a hack to monitor mouse click on the widget to
trigger a reset.

This change actually breaks im-keyboard's reset, but seems to be ok for
other because usually reset will be implemented as "clear everything".
So we still won't end up with "double commit" issue.
infamous one and we have sync release schedule.