Update cursor location when key pressed if the client window was moved
In some GTK applications, when input window is visible and the user moves
the application wdinow to another location, the cursor location will not
be updated. Then if the user presses some new keys, the input window will
not move to the new cursor location automatically. So it's necessary to
update the cursor location when key pressed if the client window was moved.